write raw XRGB32 data to temp file
I'm still not sure how "stride" fits into this. The microsoft link tells me
that the "stride" is the actual width of the buffer, and the next row is
"stride" bytes away.

But, spicy-screenshot.c seems to ignore the stride altogether, and simply
writes (width * height) bytes out to a PPM file.