Pass '-use_wallclock_as_timestamps 1' when recording raw streams
When using the 'rawvideo' codec, FFmpeg appears to set the PTS/DTS
timestamps based solely on their position in the stream (as there is no
other timing information available:

    pkt->pts = pkt->dts = pkt->pos / s->packet_size;

- https://github.com/FFmpeg/FFmpeg/blob/n3.1.7/libavformat/rawvideodec.c#L79
- http://stackoverflow.com/questions/6044330

The -use_wallclock_as_timestamps option tells FFmpeg to ignore those
timestamps and instead use the wallclock time:

    if (s->use_wallclock_as_timestamps)
        pkt->dts = pkt->pts = av_rescale_q(av_gettime(), AV_TIME_BASE_Q, st->time_base);

- https://github.com/FFmpeg/FFmpeg/blob/n3.1.7/libavformat/utils.c#L821

The goal here is to recover from any glitches in the stream, ensuring that
"real time" is always maintained as best as possible.

In my testing, it seemed to at least have no negative effect. While testing
this, I noticed #8. And it seemed to behave better with this in place.